+++ CHANGELOG.md
@@ -8,6 +8,8 @@ pull request if there was one.
 Current
 -------
 ### Added:
+- [Add Reciprocal satisfies() relationship complementing satisfiedBy() on Granularity](https://github.com/yahoo/fili/issues/222)
+
 - [CompositePhysicalTable Core Components Refactor](https://github.com/yahoo/fili/pull/179)
     * Added `ConcretePhysicalTable` and `ConcreteAvailability` to model table in druid datasource and its availabillity in the new table availability structure
     * Added class variable for `DataSourceMetadataService` and `ConfigurationLoader` into `AbstractBinderFactory` for application to access